+++ table/table_e2e_test.go
@@ -211,7 +211,7 @@ func testTable(t testing.TB) {
 		t.Logf("total consumed units: %d", atomic.LoadUint64(&totalConsumedUnits))
 	}()
 
-	ctx = meta.ListenIncomingMetadata(ctx, func(md metadata.MD) {
+	ctx = meta.WithTrailerCallback(ctx, func(md metadata.MD) {
 		atomic.AddUint64(&totalConsumedUnits, meta.ConsumedUnits(md))
 	})